Render tests should share their test data load
To speed them up.
